requirements.txt
pycryptodome
py_ecc
